Refactor Firebase initialization
Moved FirebaseInitializer from common-android to core-data module for better modularization. Updated dependencies and AndroidManifest files accordingly. Simplified Firebase usage in FirebaseRepositoryImpl by removing redundant initialization.
