+++ core/common-android/src/main/AndroidManifest.xml
@@ -28,9 +28,6 @@
                 android:name="com.keelim.commonAndroid.initialize.MobileAdsInitializer"
                 android:value="androidx.startup" />
             <meta-data
-                android:name="com.keelim.commonAndroid.initialize.FirebaseInitializer"
-                android:value="androidx.startup" />
-            <meta-data
                 android:name="com.keelim.commonAndroid.initialize.CrashInitializer"
                 android:value="androidx.startup" />
         </provider>
